+++ setup.py
@@ -58,8 +58,8 @@ SETUP_METADATA = \
     "ext_modules": [Extension("sourmash._minhash",
                                sources=["sourmash/_minhash.pyx",
                                         "third-party/smhasher/MurmurHash3.cc"],
-                               depends=["sourmash/kmer_min_hash.hh"],
-                               include_dirs=["./sourmash",
+                               depends=["include/kmer_min_hash.hh"],
+                               include_dirs=["./include",
                                              "./third-party/smhasher/"],
                                language="c++",
                                extra_compile_args=EXTRA_COMPILE_ARGS,
